privatellm/requirements.txt
langchain==0.0.177
gpt4all
chromadb==0.3.23
llama-cpp-python==0.1.50
urllib3==2.0.2
pdfminer.six==20221105
python-dotenv==1.0.0
unstructured==0.6.6
extract-msg==0.41.1
tabulate==0.9.0
pandoc==2.3
pypandoc==1.11
tqdm==4.65.0
Flask==2.1.1
constants==0.6.0
six==1.16.0
google-cloud-translate==3.11.1
google-cloud-core==2.3.2
